// actions/emissionFactorChoice/types.ts
export enum EmissionFactorChoiceTypes {
  CHOOSER_INITED = "EMISSION_FACTOR_CHOICE_CHOOSER_INITED",
  MODAL_OPEN = "EMISSION_FACTOR_CHOICE_MODAL_OPEN",
  MODAL_CLOSED = "EMISSION_FACTOR_CHOICE_MODAL_CLOSED",
  CURRENT_DATA_INITED = "EMISSION_FACTOR_CHOICE_CURRENT_DATA_INITED",
  TAG_COLUMN_OPEN = "EMISSION_FACTOR_CHOICE_TAG_COLUMN_OPEN",
  COMPUTE_METHOD_TYPE_UPDATED = "EMISSION_FACTOR_CHOICE_COMPUTE_METHOD_TYPE_UPDATED",
  COMPUTE_METHOD_ID_UPDATED = "EMISSION_FACTOR_CHOICE_COMPUTE_METHOD_ID_UPDATED",
  FILTER_DB_TOGGLED = "EMISSION_FACTOR_CHOICE_FILTER_DB_TOGGLED",
  FILTER_RECOMMENDED_TOGGLED = "EMISSION_FACTOR_CHOICE_FILTER_RECOMMENDED_TOGGLED",
  FILTER_TAG_TOGGLED = "EMISSION_FACTOR_CHOICE_FILTER_TAG_TOGGLED",
  FILTER_EF_TEXT_CHANGED = "EMISSION_FACTOR_CHOICE_FILTER_EF_TEXT_CHANGED",
  FILTER_EF_REFRESHED = "EMISSION_FACTOR_CHOICE_FILTER_EF_REFRESHED",
  LAST_CHOICE_UPDATED = "EMISSION_FACTOR_CHOICE_LAST_CHOICE_UPDATED",
  START_SEARCHING_EF = "EMISSION_FACTOR_CHOICE_START_SEARCHING_EF",
  STOP_SEARCHING_EF = "EMISSION_FACTOR_CHOICE_STOP_SEARCHING_EF",
  NOT_ENOUGH_CHARACTERS_EF_SEARCH_ERROR = "EMISSION_FACTOR_CHOICE_NOT_ENOUGH_CHARACTERS_EF_SEARCH_ERROR",
  NOT_ENOUGH_CHARACTERS_EF_SEARCH_ERROR_REMOVED = "EMISSION_FACTOR_CHOICE_NOT_ENOUGH_CHARACTERS_EF_SEARCH_ERROR_REMOVED",
}